api/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyhumps==1.6.1
simpy==4.0.1
six==1.15.0
Werkzeug==1.0.1
